+++ ext/QuantumCliffordHeckeExt/QuantumCliffordHeckeExt.jl
@@ -11,7 +11,9 @@ import Hecke: Group, GroupElem, AdditiveGroupElem,
     GroupAlgebra, GroupAlgebraElem, FqFieldElem, representation_matrix, dim, base_ring,
     multiplication_table, coefficients, abelian_group, group_algebra, rand, gens, order,
     is_commutative, FqPolyRingElem, residue_ring, coeff, zero_matrix, mod1, lift, ZZ, gen,
-    matrix, ncols, nrows, degree
+    matrix, ncols, nrows, degree, EuclideanRingResidueRingElem, quo, parent, zero, gcd,
+    polynomial_ring, characteristic, isone, mod
+import Hecke.Generic.MatSpaceElem
 import Nemo
 import Nemo: characteristic, matrix_repr, GF, ZZ, lift
 
@@ -24,12 +26,16 @@ import QECCore: AbstractECC, CSS,
 
 # exported from extension so that Documenter.jl sees them when autogenerating API lists
 export hgp, two_block_group_algebra_codes, generalized_bicycle_codes, bicycle_codes, haah_cubic_codes,
-    LPCode, LiftedCode, honeycomb_color_codes, LaCross
+    LPCode, LiftedCode, honeycomb_color_codes, LaCross, GeneralizedBicycleCode, ExtendedGeneralizedBicycleCode,
+    GeneralizedHyperGraphProductCode
 
 include("util.jl")
 include("types.jl")
 include("lifted.jl")
 include("lacross.jl")
 include("lifted_product.jl")
+include("generalized_bicycle.jl")
+include("extended_generalized_bicycle.jl")
+include("generalized_hypergraph_product.jl")
 
 end # module